+++ common/src/main/java/com/st0x0ef/stellaris/common/entities/vehicles/RocketEntity.java
@@ -1,11 +1,10 @@
 package com.st0x0ef.stellaris.common.entities.vehicles;
 
 import com.google.common.collect.Sets;
-import com.st0x0ef.stellaris.Stellaris;
 import com.st0x0ef.stellaris.client.renderers.entities.vehicle.rocket.RocketModel;
 import com.st0x0ef.stellaris.common.data.planets.Planet;
 import com.st0x0ef.stellaris.common.data_components.RocketComponent;
-import com.st0x0ef.stellaris.common.items.upgrade.RocketUpgradeItem;
+import com.st0x0ef.stellaris.common.items.RocketUpgradeItem;
 import com.st0x0ef.stellaris.common.menus.RocketMenu;
 import com.st0x0ef.stellaris.common.network.packets.SyncRocketComponentPacket;
 import com.st0x0ef.stellaris.common.registry.*;
@@ -59,7 +58,6 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 public class RocketEntity extends IVehicleEntity implements HasCustomInventoryScreen, ContainerListener {
     public int START_TIMER;
-    public boolean ROCKET_START;
     public int FUEL;
 
     public boolean needsModelChange = false;
@@ -77,6 +75,7 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
     private Player lastPlayer;
 
     private static final EntityDataAccessor<String> DATA_SKIN;
+    public static final EntityDataAccessor<Boolean> ROCKET_START = SynchedEntityData.defineId(RocketEntity.class, EntityDataSerializers.BOOLEAN);;
     private static final EntityDataAccessor<String> DATA_MODEL;
 
     public RocketEntity(EntityType<?> entityType, Level level) {
@@ -92,12 +91,10 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
         this.TANK_UPGRADE = TankUpgrade.getBasic();
 
         this.START_TIMER = 0;
-        this.ROCKET_START = false;
         this.FUEL = 0;
 
         this.currentFuelItem = ItemsRegistry.FUEL_BUCKET.get();
-        this.rocketComponent = new RocketComponent(SKIN_UPGRADE.getRocketSkinLocation().toString(), RocketModel.fromString(MODEL_UPGRADE.getModel().toString()), MODEL_UPGRADE.getMaxPlayer(), currentFuelItem.toString(), FUEL, TANK_UPGRADE.getTankCapacity());
-
+        this.rocketComponent = new RocketComponent(SKIN_UPGRADE.getRocketSkinLocation().toString(), RocketModel.fromString(MODEL_UPGRADE.getModel().toString()), currentFuelItem.toString(), FUEL, MOTOR_UPGRADE.getFluidTexture(), TANK_UPGRADE.getTankCapacity());
         this.inventory = new SimpleContainer(14);
     }
 
@@ -109,7 +106,7 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
         this.burnEntities();
         this.checkContainer();
 
-        if (ROCKET_START) {
+        if (this.entityData.get(ROCKET_START)) {
             this.spawnParticle();
             this.startTimerAndFlyMovement();
         }
@@ -161,8 +158,7 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
         compound.putString("skin", SKIN_UPGRADE.getRocketSkinLocation().toString());
         compound.putString("motor", MOTOR_UPGRADE.getFuelType().getSerializedName());
         compound.putInt("tank", TANK_UPGRADE.getTankCapacity());
-        compound.putInt("maxPlayer", MODEL_UPGRADE.getMaxPlayer());
-
+        compound.putBoolean("rocketStart", this.entityData.get(ROCKET_START));
     }
 
     @Override
@@ -186,9 +182,9 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
             }
         }
 
-        this.MODEL_UPGRADE = new ModelUpgrade(RocketModel.fromString(compound.getString("model")), compound.getInt("maxPlayer"));
+        this.MODEL_UPGRADE = new ModelUpgrade(RocketModel.fromString(compound.getString("model")));
         this.SKIN_UPGRADE = new SkinUpgrade(ResourceLocation.parse(compound.getString("skin")));
-        this.MOTOR_UPGRADE = new MotorUpgrade(FuelType.Type.fromString(compound.getString("motor")));
+        this.MOTOR_UPGRADE = new MotorUpgrade(FuelType.Type.fromString(compound.getString("motor")), ResourceLocation.parse(compound.getString("fuel_texture")));
         this.TANK_UPGRADE = new TankUpgrade(compound.getInt("tank"));
     }
 
@@ -207,8 +203,8 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
         this.entityData.set(DATA_SKIN, skinUpgrade.getRocketSkinLocation().toString());
     }
 
-    public ModelUpgrade getModelData() { // not sure if this is needed yet
-        return new ModelUpgrade(RocketModel.fromString(this.entityData.get(DATA_MODEL)), 1);
+    public ModelUpgrade getModelData() {
+        return new ModelUpgrade(RocketModel.fromString(this.entityData.get(DATA_MODEL)));
     }
 
     public void setModelData(ModelUpgrade modelUpgrade) {
@@ -239,6 +235,9 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
             if (player.isCrouching()) {
                 if (!tryFillUpRocket(player.getMainHandItem().getItem())) {
                     this.openCustomInventoryScreen(player);
+                } else {
+                    player.getItemInHand(hand).grow(-1);
+                    player.getInventory().add(new ItemStack(Items.BUCKET));
                 }
                 return InteractionResult.CONSUME;
             }
@@ -254,11 +253,9 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
     }
 
     public boolean canPlayerRide() {
-        int maxPlayer = this.MODEL_UPGRADE.getMaxPlayer();
+        int maxPlayer = this.MODEL_UPGRADE.getModel().getMaxPlayerNumber();
         return this.getPassengers().size() < maxPlayer;
     }
-
-
 
     @Override
     public Vec3 getPassengerRidingPosition(Entity entity) {
@@ -387,8 +384,8 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
 
         if (player != null) {
             if (this.FUEL > 0 || player.isCreative()) {
-                if (!this.ROCKET_START) {
-                    this.ROCKET_START = true;
+                if (!this.entityData.get(ROCKET_START)) {
+                    this.entityData.set(ROCKET_START, true);
                     this.level().playSound(null, this, SoundRegistry.ROCKET_SOUND.get(), SoundSource.NEUTRAL, 1, 1);
                 }
             } else {
@@ -447,7 +444,7 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
     }
     public ItemStack getRocketItem() {
         ItemStack itemStack = new ItemStack(ItemsRegistry.ROCKET.get(), 1);
-        rocketComponent = new RocketComponent(SkinUpgrade.getBasic().getNameSpace(), ModelUpgrade.getBasic().getModel(), MODEL_UPGRADE.getMaxPlayer(), currentFuelItem.toString(), FUEL, TANK_UPGRADE.getTankCapacity());
+        rocketComponent = new RocketComponent(SkinUpgrade.getBasic().getNameSpace(), ModelUpgrade.getBasic().getModel(), currentFuelItem.toString(), FUEL, MOTOR_UPGRADE.getFluidTexture(), TANK_UPGRADE.getTankCapacity());
         itemStack.set(DataComponentsRegistry.ROCKET_COMPONENT.get(), rocketComponent);
 
         return itemStack;
@@ -633,8 +630,7 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
     }
 
     public void syncRocketData(ServerPlayer player) {
-        this.rocketComponent = new RocketComponent(SKIN_UPGRADE.getRocketSkinLocation().toString(), RocketModel.fromString(MODEL_UPGRADE.getModel().toString()), MODEL_UPGRADE.getMaxPlayer(), currentFuelItem.toString(), FUEL, TANK_UPGRADE.getTankCapacity());
-
+        this.rocketComponent = new RocketComponent(SKIN_UPGRADE.getRocketSkinLocation().toString(), RocketModel.fromString(MODEL_UPGRADE.getModel().toString()), currentFuelItem.toString(), FUEL, MOTOR_UPGRADE.getFluidTexture(), TANK_UPGRADE.getTankCapacity());
         if (!level().isClientSide()) {
             NetworkManager.sendToPlayer(player, new SyncRocketComponentPacket(rocketComponent));
         }
@@ -679,7 +675,6 @@ public class RocketEntity extends IVehicleEntity implements HasCustomInventorySc
     public Packet<ClientGamePacketListener> getAddEntityPacket(ServerEntity entity) {
         return NetworkManager.createAddEntityPacket(this, entity);
     }
-
 
     static {
         DATA_SKIN = SynchedEntityData.defineId(RocketEntity.class, EntityDataSerializers.STRING);