+++ common/src/main/java/com/st0x0ef/stellaris/common/rocket_upgrade/ModelUpgrade.java
@@ -4,22 +4,16 @@ import com.st0x0ef.stellaris.client.renderers.entities.vehicle.rocket.RocketMode
 
 public class ModelUpgrade extends RocketUpgrade {
     final private RocketModel model;
-    final private int maxPlayer;
 
-    public ModelUpgrade(RocketModel model, int maxPlayer) {
+    public ModelUpgrade(RocketModel model) {
         this.model = model;
-        this.maxPlayer = maxPlayer;
     }
 
     public RocketModel getModel() {
         return model;
     }
 
-    public int getMaxPlayer() {
-        return maxPlayer;
-    }
-
     public static ModelUpgrade getBasic() {
-        return new ModelUpgrade(RocketModel.NORMAL, 2);
+        return new ModelUpgrade(RocketModel.TINY);
     }
 }